Rename errors package
