+++ x/profile/keeper/msg_server_update_coordinator_address.go
@@ -6,7 +6,7 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/tendermint/spn/pkg/critical"
+	spnerrors "github.com/tendermint/spn/pkg/errors"
 	"github.com/tendermint/spn/x/profile/types"
 )
 
@@ -32,7 +32,7 @@ func (k msgServer) UpdateCoordinatorAddress(
 
 	if !k.HasCoordinator(ctx, coordByAddress.CoordinatorId) {
 		return &types.MsgUpdateCoordinatorAddressResponse{},
-			critical.Error("a coordinator address is associated to a non-existent coordinator ID")
+			spnerrors.Critical("a coordinator address is associated to a non-existent coordinator ID")
 	}
 
 	coord := k.GetCoordinator(ctx, coordByAddress.CoordinatorId)
